+++ src/routes/Morph/index.js
@@ -7,6 +7,7 @@ import Tool from '../../components/Tool'
 
 import MorphForm from './MorphForm'
 import MorphResults from './MorphResults'
+import morphService from './MorphService'
 import styles from './styles'
 
 export const morphInfo = {
@@ -16,13 +17,27 @@ export const morphInfo = {
     'is a tool for modeling historical sound change according to arbitrary rules of phonetic change.'
 }
 
-const Morph = props => (
-  <Tool toolInfo={morphInfo}>
-    <Notice>This tool is still in development.</Notice>
-    <MorphForm styles={props.classes} />
-    <MorphResults styles={props.classes} />
-  </Tool>
-)
+class Morph extends React.Component {
+  constructor (props) {
+    super(props)
+    this.onChange = this.onChange.bind(this)
+    this.state = morphService.getData()
+  }
+
+  onChange (e) {
+    return
+  }
+
+  render () {
+    return (
+      <Tool toolInfo={morphInfo}>
+        <Notice>This tool is still in development.</Notice>
+        <MorphForm styles={this.props.classes} />
+        <MorphResults styles={this.props.classes} />
+      </Tool>
+    )
+  }
+}
 
 Morph.propTypes = {
   classes: PropTypes.object